♻️ refactor(messageHandler): split generateReply into separate module
- move generateReply function to src/handlers/replyGenerator.ts for clarity
- relocate discord helper functions to src/utils/discordHelpers.ts
- update imports and references in messageHandler accordingly

♻️ refactor(handlers): rename createMessage.ts to messageHandler.ts

- rename file to better reflect message handling role
- remove unused imports and simplify code in messageHandler.ts

💄 style(utils): add discordHelpers.ts for discord text utilities

- implement functions to fix mentions, replace emoji shortcodes
- add createChatMessage and summariseConversation helpers
